ISSUE #4625 - Moved virtual list to separate file, fixed ui quirk in list
